requirements.txt
tensorflow
numpy
scikit-learn
gzip
struct
pillow
matplotlib
